+++ src/main/java/io/pakland/mdas/githubstats/application/exceptions/OrganizationNotFound.java
@@ -1,7 +1,7 @@
 package io.pakland.mdas.githubstats.application.exceptions;
 
 public class OrganizationNotFound extends Exception {
-    public OrganizationNotFound(Long id) {
+    public OrganizationNotFound(Integer id) {
         super("Organization with id: " + id + " not found");
     }
 }
